WIP: moved linter code to linter/
