+++ linter/models/rbc_dtcc_iid.yaml
@@ -1,0 +1,86 @@
+name: Real Business Cycle
+
+model_type: dtcc
+
+symbols:
+
+   exogenous: [e_z]
+   states: [z, k]
+   controls: [n, i]
+   strange_params: [i,i]
+   expectations: [m]
+   values: [V]
+   parameters: [beta, sigma, beta, eta, chi, delta, alpha, rho, zbar, sig_z]
+   rewards: [u]
+
+definitions:
+    y: exp(z)*k^alpha*n^(1-alpha)
+    c: y - i
+    rk: alpha*y/k
+    w: (1-alpha)*y/n
+
+equations:
+
+    arbitrage:
+        - chi*n^eta*c^sigma - w                      | 0.0 <= n <= inf
+        - 1 - beta*(c/c(1))^(sigma)*(1-delta+rk(1))  | 0.0 <= i <= inf
+
+
+    transition:
+        - z = rho*z(-1) + e_z
+        - k = (1-delta)*k(-1) + i(-1)
+
+    value:
+        - V = c^(1-sigma)/(1-sigma) - chi*n^(1+eta)/(1+eta) + beta*V(1)
+
+    felicity:
+        - u =  c^(1-sigma)/(1-sigma) - chi*n^(1+eta)/(1+eta)
+
+    expectation:
+        - m = beta/c(1)^sigma*(1-delta+rk(1))
+
+    direct_response:
+        - n = ((1-alpha)*exp(z)*k^alpha*m/chi)^(1/(eta+alpha))
+        - i = exp(z)*k^alpha*n^(1-alpha) - (m)^(-1/sigma)
+
+calibration:
+
+    # parameters
+    beta : 0.99
+    phi: 1
+    delta : 0.025
+    alpha : 0.33
+    rho : 0.8
+    sigma: 5
+    eta: 1
+    sig_z: 0.016
+    zbar: 0.0
+    chi : w/c^sigma/n^eta
+    c_i: 1.5
+    c_y: 0.5
+    e_z: 0.0
+
+    # endogenous variables
+    z: 0.0
+    n: 0.33
+    z: zbar
+    rk: 1/beta-1+delta
+    w: (1-alpha)*exp(z)*(k/n)^(alpha)
+    k: n/(rk/alpha)^(1/(1-alpha))
+    y: exp(z)*k^alpha*n^(1-alpha)
+    i: delta*k
+    c: y - i
+    V: log(c)/(1-beta)
+    u: c^(1-sigma)/(1-sigma) - chi*n^(1+eta)/(1+eta)
+    m: beta/c^sigma*(1-delta+rk)
+
+domain:
+    z: [-2*sig_z/(1-rho^2)^0.5,  2*sig_z/(1-rho^2)^0.5]
+    k: [k*0.5, k*1.5]
+
+exogenous: !Normal
+    Sigma: sig_z^2
+
+options:
+    grid: !Cartesian
+        orders: [5, 50]
